Boundary Conditions

ContinuousWavelets.WaveletBoundary — Type

the abstract type for the various types of boundaries

ContinuousWavelets.PerBoundary — Type

PerBoundary() <: WaveletBoundary

standard periodic boundary assumption made by the fft. Aliases of NaivePer and Periodic.

ContinuousWavelets.ZPBoundary — Type

ZPBoundary() <: WaveletBoundary

zero pads the signal before doing an fft, rounding up to the nearest power of two. Alias of padded.

ContinuousWavelets.SymBoundary — Type

SymBoundary() <: WaveletBoundary

symmetric boundary, as in the DCT type II. Repeats the edge value, which alleviates derivative discontinuities. Alias of DEFAULT_BOUNDARY and SymBound.